docs: add impact section to feature template
As discussed in retro, add a section to describe the impact the proposed changes will have on the behaviour of our apps and services.

For changes with a significant chance of unintended side effects we can add a risk level as well. We can use this risk level as a guide to ensure changes are tested properly and rollback provisions are made if needed.